Add App
App contains context, dispatcher and error handler.

Dispatcher now contains only a Vec of handlers.
Also it implements Handler trait and we can use dispatcher in predicates.

DispatcherBuilder exists because Dispatcher contains Arc<Vec<_>> in order
to clone handlers list cheaply. Alternatively we need to use a Mutex and it is not a best decision.

InputHandler trait used because Handler will require Clone and it ruines object safety.